more restructuring for FMA
