+++ app/build.gradle
@@ -17,15 +17,11 @@ apply plugin: 'android-unit-test'
 
 repositories {
     mavenCentral()
-    flatDir {
-        dirs 'libs'
-    }
 }
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile 'com.android.support:appcompat-v7:21.0.3'
-    compile 'com.spotify.sdk:spotifysdk:1.0.0-beta6@aar'
     testCompile 'junit:junit:4.10'
     testCompile 'org.robolectric:robolectric:2.4'
 }
