+++ app/src/main/res/layout/activity_player.xml
@@ -1,5 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent">
+
+    <LinearLayout
         android:id="@+id/player_layout"
         android:orientation="vertical"
         android:layout_width="match_parent"
@@ -136,3 +140,14 @@
         </LinearLayout>
 
     </LinearLayout>
+
+    <ProgressBar
+        android:id="@+id/progbar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        android:indeterminateOnly="true"
+        android:visibility="gone"
+        style="@android:style/Widget.ProgressBar.Large" />
+
+</RelativeLayout>
